+++ app/components/common/guards/hidden-navigation.directive.ts
