+++ app/components/common/guards/navigation-guard.component.ts
