+++ src/@batch-flask/ui/editor/editor.module.ts
@@ -1,12 +1,13 @@
 import { NgModule } from "@angular/core";
 
 import { EditorComponent } from "./editor.component";
+import { MonacoLoader } from "./monaco-loader.service";
 
 const components = [EditorComponent];
 @NgModule({
     declarations: components,
     imports: [],
-    providers: [],
+    providers: [MonacoLoader],
     exports: components,
 })
 export class EditorModule {
